requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
openai==1.35.10
pywebview==5.2
jinja2==3.1.4
anthropic==0.34.2
google-generativeai==0.7.2
ollama==0.3.2
